chore(graphics): structure post processing code
