+++ src/plugins/graphics/src/components.rs
@@ -3,6 +3,7 @@ pub(crate) mod child_meshes;
 pub(crate) mod effect_material_handle;
 pub(crate) mod material_override;
 pub(crate) mod model_render_layers;
+pub(crate) mod post_process_camera;
 
 #[cfg(not(feature = "debug-utils"))]
 pub mod no_debug_cam;
