+++ src/plugins/graphics/src/lib.rs
@@ -11,45 +11,23 @@ use crate::{
 		camera_labels::OutlinePass,
 		child_meshes::ChildMeshOf,
 		model_render_layers::ModelRenderLayers,
+		post_process_camera::PostProcessCamera,
 	},
 	materials::effect_material::EffectMaterial,
 	observers::insert_render_target::InsertRenderTarget,
+	resources::post_process_pipeline::{PostProcessLabel, PostProcessNode, PostProcessPipeline},
 	system_params::highlight::{HighlightParam, HighlightParamMut},
 };
 use bevy::{
-	core_pipeline::{
-		FullscreenShader,
-		core_3d::graph::{Core3d, Node3d},
-	},
-	ecs::query::QueryItem,
+	core_pipeline::core_3d::graph::Core3d,
 	prelude::*,
 	render::{
 		RenderApp,
 		RenderStartup,
-		extract_component::{
-			ComponentUniforms,
-			DynamicUniformIndex,
-			ExtractComponent,
-			ExtractComponentPlugin,
-			UniformComponentPlugin,
-		},
+		extract_component::{ExtractComponentPlugin, UniformComponentPlugin},
 		extract_resource::ExtractResourcePlugin,
-		render_asset::RenderAssets,
-		render_graph::{
-			NodeRunError,
-			RenderGraphContext,
-			RenderGraphExt,
-			RenderLabel,
-			ViewNode,
-			ViewNodeRunner,
-		},
-		render_resource::{
-			binding_types::{sampler, texture_2d, uniform_buffer},
-			*,
-		},
-		renderer::{RenderContext, RenderDevice},
-		texture::GpuImage,
-		view::ViewTarget,
+		render_graph::{RenderGraphExt, ViewNodeRunner},
+		render_resource::*,
 	},
 };
 use common::{
@@ -157,8 +135,8 @@ where
 
 		app.init_resource::<WindowSize>()
 			.add_plugins(ExtractResourcePlugin::<CameraRenderTarget<OutlinePass>>::default())
-			.add_plugins(ExtractComponentPlugin::<OutlineSettings>::default())
-			.add_plugins(UniformComponentPlugin::<OutlineSettings>::default())
+			.add_plugins(ExtractComponentPlugin::<PostProcessCamera>::default())
+			.add_plugins(UniformComponentPlugin::<PostProcessCamera>::default())
 			.register_required_components_with::<Ui, TDebugCam>(self.debug_cam)
 			.add_observer(WorldPass::insert_render_target)
 			.add_observer(OutlinePass::insert_render_target)
@@ -181,16 +159,9 @@ where
 			);
 
 		app.sub_app_mut(RenderApp)
-			.add_systems(RenderStartup, init_post_process_pipeline)
+			.add_systems(RenderStartup, PostProcessPipeline::init)
 			.add_render_graph_node::<ViewNodeRunner<PostProcessNode>>(Core3d, PostProcessLabel)
-			.add_render_graph_edges(
-				Core3d,
-				(
-					Node3d::Tonemapping,
-					PostProcessLabel,
-					Node3d::EndMainPassPostProcessing,
-				),
-			);
+			.add_render_graph_edges(Core3d, PostProcessLabel::EDGES);
 	}
 }
 
@@ -245,154 +216,4 @@ impl<TDebugCam, TDependencies> WorldCameras for GraphicsPlugin<TDebugCam, TDepen
 impl<TDebugCam, TDependencies> HandlesGraphics for GraphicsPlugin<TDebugCam, TDependencies> {
 	type THighlight = HighlightParam<'static, 'static>;
 	type THighlightMut = HighlightParamMut<'static, 'static>;
-}
-
-// FIXME:: MOVE STUFF BELLOW TO PROPER MODULES
-
-const SHADER_ASSET_PATH: &str = "shaders/post_processing.wgsl";
-
-#[derive(Component, Default, Clone, Copy, ExtractComponent, ShaderType)]
-struct OutlineSettings {
-	color: LinearRgba,
-}
-
-#[derive(Resource)]
-struct PostProcessPipeline {
-	layout: BindGroupLayoutDescriptor,
-	sampler: Sampler,
-	pipeline_id: CachedRenderPipelineId,
-}
-
-#[derive(Debug, Hash, PartialEq, Eq, Clone, RenderLabel)]
-struct PostProcessLabel;
-
-#[derive(Default)]
-struct PostProcessNode;
-
-impl ViewNode for PostProcessNode {
-	type ViewQuery = (
-		&'static ViewTarget,
-		&'static DynamicUniformIndex<OutlineSettings>,
-	);
-
-	fn run(
-		&self,
-		_: &mut RenderGraphContext,
-		render_context: &mut RenderContext,
-		(view_target, settings_index): QueryItem<Self::ViewQuery>,
-		world: &World,
-	) -> Result<(), NodeRunError> {
-		// Get render pipeline
-		let Some(post_process_pipeline) = world.get_resource::<PostProcessPipeline>() else {
-			// FIXME: ERROR?
-			return Ok(());
-		};
-		let Some(cache) = world.get_resource::<PipelineCache>() else {
-			// FIXME: ERROR?
-			return Ok(());
-		};
-		let Some(pipeline) = cache.get_render_pipeline(post_process_pipeline.pipeline_id) else {
-			// FIXME: ERROR?
-			return Ok(());
-		};
-
-		// get `PostProcessSettings` as bindings
-		let Some(settings) = world.get_resource::<ComponentUniforms<OutlineSettings>>() else {
-			// FIXME: ERROR?
-			return Ok(());
-		};
-		let Some(settings_binding) = settings.uniforms().binding() else {
-			// FIXME: ERROR?
-			return Ok(());
-		};
-
-		// get outline target texture
-		let Some(gpu_images) = world.get_resource::<RenderAssets<GpuImage>>() else {
-			// FIXME: ERROR?
-			return Ok(());
-		};
-		let Some(outline) = world.get_resource::<CameraRenderTarget<OutlinePass>>() else {
-			// FIXME: ERROR?
-			return Ok(());
-		};
-		let Some(outline_gpu) = gpu_images.get(&outline.handle) else {
-			// FIXME: ERROR?
-			return Ok(());
-		};
-
-		let post_process = view_target.post_process_write();
-		let bind_group = render_context.render_device().create_bind_group(
-			"post_process_bind_group",
-			&cache.get_bind_group_layout(&post_process_pipeline.layout),
-			&BindGroupEntries::sequential((
-				post_process.source,
-				&post_process_pipeline.sampler,
-				&outline_gpu.texture_view,
-				&outline_gpu.sampler,
-				settings_binding.clone(),
-			)),
-		);
-
-		let mut render_pass = render_context.begin_tracked_render_pass(RenderPassDescriptor {
-			label: Some("post_process_pass"),
-			color_attachments: &[Some(RenderPassColorAttachment {
-				view: post_process.destination,
-				depth_slice: None,
-				resolve_target: None,
-				ops: Operations::default(),
-			})],
-			depth_stencil_attachment: None,
-			timestamp_writes: None,
-			occlusion_query_set: None,
-		});
-
-		render_pass.set_render_pipeline(pipeline);
-		render_pass.set_bind_group(0, &bind_group, &[settings_index.index()]);
-		render_pass.draw(0..3, 0..1);
-
-		Ok(())
-	}
-}
-
-fn init_post_process_pipeline(
-	mut commands: Commands,
-	render_device: Res<RenderDevice>,
-	asset_server: Res<AssetServer>,
-	fullscreen_shader: Res<FullscreenShader>,
-	pipeline_cache: Res<PipelineCache>,
-) {
-	let layout = BindGroupLayoutDescriptor::new(
-		"post_process_bind_group_layout",
-		&BindGroupLayoutEntries::sequential(
-			ShaderStages::FRAGMENT,
-			(
-				texture_2d(TextureSampleType::Float { filterable: true }),
-				sampler(SamplerBindingType::Filtering),
-				texture_2d(TextureSampleType::Float { filterable: true }),
-				sampler(SamplerBindingType::Filtering),
-				uniform_buffer::<OutlineSettings>(true),
-			),
-		),
-	);
-	let pipeline_id = pipeline_cache.queue_render_pipeline(RenderPipelineDescriptor {
-		label: Some("post_process_pipeline".into()),
-		layout: vec![layout.clone()],
-		vertex: fullscreen_shader.to_vertex_state(),
-		fragment: Some(FragmentState {
-			shader: asset_server.load(SHADER_ASSET_PATH),
-			targets: vec![Some(ColorTargetState {
-				format: TextureFormat::Rgba16Float,
-				blend: None,
-				write_mask: ColorWrites::ALL,
-			})],
-			..default()
-		}),
-		..default()
-	});
-
-	commands.insert_resource(PostProcessPipeline {
-		layout,
-		sampler: render_device.create_sampler(&SamplerDescriptor::default()),
-		pipeline_id,
-	});
 }
