+++ src/plugins/graphics/src/resources.rs
@@ -1,2 +1,3 @@
 pub(crate) mod camera_render_target;
+pub(crate) mod post_process_pipeline;
 pub(crate) mod window_size;
