+++ src/plugins/graphics/src/systems.rs
@@ -1,3 +1,4 @@
+pub(crate) mod init_post_process_pipeline;
 pub(crate) mod modify_material;
 pub(crate) mod no_waiting_pipelines;
 pub(crate) mod populate_missing_model_render_layers;
